+++ projects/w11k/tydux/src/lib/commands-mutators.ts
@@ -1,7 +1,8 @@
 import {Commands, CommandsStateType} from "./commands";
 
-export function assignFieldCommand<T extends Commands<any>, F extends keyof CommandsStateType<T>>(
-    commands: T, field: F): (value: CommandsStateType<T>[F]) => void {
+export function createAssignFieldCommand<T extends Commands<any>, F extends keyof CommandsStateType<T>>(
+    commands: T, field: F
+): (value: CommandsStateType<T>[F]) => void {
 
     return function (this: any, value: CommandsStateType<T>[F]) {
         (this as any).state = {