+++ projects/w11k/tydux/src/public_api.ts
@@ -8,7 +8,7 @@ export * from "./lib/development";
 export * from "./lib/store";
 export * from "./lib/store-global";
 export * from "./lib/utils";
-export * from "./lib/ordered-map.facade";
+export * from "./lib/ordered-map";
 export * from "./lib/commands-mutators";
 
 // TODO separate entrypoint for testing support