+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 13.5.2
+
+- changed OrderedMapFacade to command mutator
+- assignFieldCommand -> createAssignFieldCommand
+
 # Version 13.5.0
 
 - added mutator: assignFieldCommand